feat(command-input): remove platform detection, and fix SSR hydration mismatch
