+++ .changeset/witty-eagles-brush.md
@@ -1,0 +1,5 @@
+---
+'@fidely-ui/react': minor
+---
+
+`feat(command-input)`: remove platform detection, and fix SSR hydration mismatch
