--- packages/fidely-ui/src/utils/is-mobile-device.ts
@@ -1,6 +1,0 @@
-export function isMobileDevice() {
-  if (typeof navigator === 'undefined') return false
-  return /android|bb\d+|meego|avantgo|bada\/|blackberry|blazer|compal|elaine|fennec|hiptop|iemobile|ip(?:hone|od|pad)|iris|kindle|lge |maemo|midp|mmp|mobile.+firefox|netfront|opera m(?:ob|in)i|palm(?: os)?|phone|p(?:ixi|re)\/|plucker|pocket|psp|series(?:4|6)0|symbian|treo|up\.(?:browser|link)|vodafone|wap|windows ce|windows phone|xda|xiino|zte-/i.test(
-    navigator.userAgent
-  )
-}
